// src/app/series-rm/interfaces/allEpisodes.ts
export interface ResponseAllEpisodes {
    info:    Info;
    results: Episode[];
}

export interface Info {
    count: number;
    pages: number;
    next:  null;
    prev:  string;
}

export interface Episode {
    id:         number;
    name:       string;
    air_date:   string;
    episode:    string;
    characters: string[];
    url:        string;
    created:    Date;
}